requirements.txt
# Core HTTP client for API communication
aiohttp>=3.8.0

# Data visualization and analysis
matplotlib>=3.5.0
pandas>=1.3.0
numpy>=1.20.0

# Azure services (optional, for cloud storage)
azure-identity>=1.10.0
azure-keyvault-secrets>=4.4.0

# Encryption and security
cryptography>=37.0.0

# Testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.18.0

# Documentation (optional)
sphinx>=4.0.0

# Web interface
flask>=2.0.0